Use correct tmp var when iterating over class hierarchy #12776
