+++ extensions/panache/hibernate-orm-rest-data-panache/deployment/src/test/java/io/quarkus/hibernate/orm/rest/data/panache/deployment/entity/AbstractItem.java
@@ -1,19 +1,11 @@
 package io.quarkus.hibernate.orm.rest.data.panache.deployment.entity;
 
 import javax.json.bind.annotation.JsonbTransient;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
 import javax.persistence.ManyToOne;
 import javax.persistence.MappedSuperclass;
 
-import io.quarkus.hibernate.orm.panache.PanacheEntityBase;
-
 @MappedSuperclass
-public abstract class AbstractItem<IdType> extends PanacheEntityBase {
-
-    @Id
-    @GeneratedValue
-    public IdType id;
+public abstract class AbstractItem<IdType extends Number> extends AbstractEntity<IdType> {
 
     public String name;
