+++ extensions/panache/hibernate-orm-rest-data-panache/deployment/src/test/java/io/quarkus/hibernate/orm/rest/data/panache/deployment/entity/PanacheEntityResourceDeleteMethodTest.java
@@ -12,8 +12,8 @@ class PanacheEntityResourceDeleteMethodTest extends AbstractDeleteMethodTest {
     @RegisterExtension
     static final QuarkusUnitTest TEST = new QuarkusUnitTest()
             .setArchiveProducer(() -> ShrinkWrap.create(JavaArchive.class)
-                    .addClasses(Collection.class, CollectionsController.class, AbstractItem.class, Item.class,
-                            ItemsController.class)
+                    .addClasses(Collection.class, CollectionsController.class, AbstractEntity.class, AbstractItem.class,
+                            Item.class, ItemsController.class)
                     .addAsResource("application.properties")
                     .addAsResource("import.sql"));
 }
