+++ extensions/panache/hibernate-orm-rest-data-panache/deployment/src/test/java/io/quarkus/hibernate/orm/rest/data/panache/deployment/repository/AbstractEntity.java
@@ -1,0 +1,21 @@
+package io.quarkus.hibernate.orm.rest.data.panache.deployment.repository;
+
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.MappedSuperclass;
+
+@MappedSuperclass
+public abstract class AbstractEntity<IdType extends Number> {
+
+    @Id
+    @GeneratedValue
+    private IdType id;
+
+    public IdType getId() {
+        return id;
+    }
+
+    public void setId(IdType id) {
+        this.id = id;
+    }
+}
