+++ extensions/panache/hibernate-orm-rest-data-panache/deployment/src/test/java/io/quarkus/hibernate/orm/rest/data/panache/deployment/repository/AbstractItem.java
@@ -1,30 +1,16 @@
 package io.quarkus.hibernate.orm.rest.data.panache.deployment.repository;
 
 import javax.json.bind.annotation.JsonbTransient;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
 import javax.persistence.ManyToOne;
 import javax.persistence.MappedSuperclass;
 
 @MappedSuperclass
-public abstract class AbstractItem<IdType> {
-
-    @Id
-    @GeneratedValue
-    private IdType id;
+public abstract class AbstractItem<IdType extends Number> extends AbstractEntity<IdType> {
 
     private String name;
 
     @ManyToOne(optional = false)
     private Collection collection;
-
-    public IdType getId() {
-        return id;
-    }
-
-    public void setId(IdType id) {
-        this.id = id;
-    }
 
     public String getName() {
         return name;
