+++ extensions/panache/hibernate-orm-rest-data-panache/deployment/src/test/java/io/quarkus/hibernate/orm/rest/data/panache/deployment/repository/PanacheRepositoryResourceDeleteMethodTest.java
@@ -12,8 +12,9 @@ class PanacheRepositoryResourceDeleteMethodTest extends AbstractDeleteMethodTest
     @RegisterExtension
     static final QuarkusUnitTest TEST = new QuarkusUnitTest()
             .setArchiveProducer(() -> ShrinkWrap.create(JavaArchive.class)
-                    .addClasses(Collection.class, CollectionsController.class, CollectionsRepository.class, AbstractItem.class,
-                            Item.class, ItemsController.class, ItemsRepository.class)
+                    .addClasses(Collection.class, CollectionsController.class, CollectionsRepository.class,
+                            AbstractEntity.class, AbstractItem.class, Item.class, ItemsController.class,
+                            ItemsRepository.class)
                     .addAsResource("application.properties")
                     .addAsResource("import.sql"));
 }
