+++ extensions/panache/hibernate-orm-rest-data-panache/deployment/src/test/java/io/quarkus/hibernate/orm/rest/data/panache/deployment/repository/PanacheRepositoryResourceDevModeTest.java
@@ -12,7 +12,8 @@ public class PanacheRepositoryResourceDevModeTest extends AbstractDevModeTest {
     @RegisterExtension
     static final QuarkusDevModeTest TEST = new QuarkusDevModeTest()
             .setArchiveProducer(() -> ShrinkWrap.create(JavaArchive.class)
-                    .addClasses(Collection.class, AbstractItem.class, Item.class, ItemsController.class, ItemsRepository.class)
+                    .addClasses(Collection.class, AbstractEntity.class, AbstractItem.class, Item.class,
+                            ItemsController.class, ItemsRepository.class)
                     .addAsResource("application.properties")
                     .addAsResource("import.sql"));
